+++ scripts/modules/constants.py
@@ -3,7 +3,7 @@
 from enum import Enum
 
 
-SCRIPT_VERSION = "2.6"
+SCRIPT_VERSION = "2.7"
 
 
 class ExitCode(Enum):